# lib/api-spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
paths:
...
  /reports:
    get:
      tags:
      - reports
      summary: List Reports
      operationId: getReports
      parameters:
      - name: studyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
      - name: reportType
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Reporttype
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportOut'
                title: Response Getreports
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - reports
      summary: Create Report
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/dqa-daily:
    post:
      tags:
      - reports
      summary: Create Dqa Daily
      operationId: createDqaDailyReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDqaDailyInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/dqa-final:
    post:
      tags:
      - reports
      summary: Create Dqa Final
      operationId: createDqaFinalReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDqaFinalInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{report_id}:
    get:
      tags:
      - reports
      summary: Get Report
      operationId: getReport
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - reports
      summary: Delete Report
      operationId: deleteReport
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{report_id}/preview:
    get:
      tags:
      - reports
      summary: Preview Report
      operationId: previewReport
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{report_id}/download:
    get:
      tags:
      - reports
      summary: Download Report
      operationId: downloadReport
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          title: Report Id
      - name: format
        in: query
        required: false
        schema:
          type: string
          pattern: ^(pdf|docx)$
          default: pdf
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
            application/pdf: {}
            application/vnd.openxmlformats-officedocument.wordprocessingml.document: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{report_id}/generate:
    post:
      tags:
      - reports
      summary: Generate Report
      operationId: generateReport
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{report_id}/share:
    post:
      tags:
      - reports
      summary: Share Report
      operationId: shareReport
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          title: Report Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareReportInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GenerateDqaDailyInput:
      properties:
        studyId:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
        reportDate:
          anyOf:
          - type: string
          - type: 'null'
          title: Reportdate
        sendEmail:
          type: boolean
          title: Sendemail
          default: false
        runAi:
          type: boolean
          title: Runai
          default: true
      type: object
      title: GenerateDqaDailyInput
    GenerateDqaFinalInput:
      properties:
        studyId:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
        runAi:
          type: boolean
          title: Runai
          default: true
        sendEmail:
          type: boolean
          title: Sendemail
          default: false
      type: object
      title: GenerateDqaFinalInput
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OkResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
      type: object
      title: OkResponse
...
    ReportInput:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
          default: ''
        format:
          type: string
          title: Format
          default: pdf
        reportType:
          type: string
          title: Reporttype
          default: custom
        studyId:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
        reportDate:
          anyOf:
          - type: string
          - type: 'null'
          title: Reportdate
        promptId:
          anyOf:
          - type: string
          - type: 'null'
          title: Promptid
        projectIds:
          items:
            type: string
          type: array
          title: Projectids
          default: []
      type: object
      required:
      - title
      title: ReportInput
    ReportOut:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
        format:
          type: string
          title: Format
        reportType:
          type: string
          title: Reporttype
          default: custom
        studyId:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
        reportDate:
          anyOf:
          - type: string
          - type: 'null'
          title: Reportdate
        promptId:
          anyOf:
          - type: string
          - type: 'null'
          title: Promptid
        promptName:
          anyOf:
          - type: string
          - type: 'null'
          title: Promptname
        projectIds:
          items:
            type: string
          type: array
          title: Projectids
        projectNames:
          items:
            type: string
          type: array
          title: Projectnames
        generatedContent:
          anyOf:
          - type: string
          - type: 'null'
          title: Generatedcontent
        downloadUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Downloadurl
        pageCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pagecount
        fileSizeKb:
          anyOf:
          - type: number
          - type: 'null'
          title: Filesizekb
        generatedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Generatedat
        createdAt:
          type: string
          title: Createdat
      type: object
      required:
      - id
      - title
      - description
      - status
      - format
      - projectIds
      - projectNames
      - createdAt
      title: ReportOut
...
    ShareReportInput:
      properties:
        recipients:
          items:
            type: string
          type: array
          title: Recipients
        subject:
          type: string
          title: Subject
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
      type: object
      required:
      - recipients
      - subject
      title: ShareReportInput
    ShareResult:
      properties:
        success:
          type: boolean
          title: Success
        recipientsCount:
          type: integer
          title: Recipientscount
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - recipientsCount
      - message
      title: ShareResult
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: /api
  description: Base API path